requirements.txt
playwright==1.48.0
requests==2.31.0
openpyxl==3.1.2
openai>=1.0.0
python-dotenv>=1.0.0
